Epoch: Creating an epoch now takes references on buffers + versions
When an epoch is created using the constructor Epoch(Structure, Buffer),
it will call take_reference() on both.

This was necessary to ensure that the destructor doesn't fail, as it
releases references and fails if the refcnt is 0. It releases the user
of the object from the burden of manually taking references in this
situation.