DynamicExtension: fixed some Epoch-related bugs
The epochs must be released in the destructor prior to releasing the
buffers and structures, as otherwise there are references remaining to
these objects and their destructors will fail.

Additionally, fixed a bug in the constructor resulting in a memory leak
due to allocating an extra starting version and buffer.