FIFOScheduler: added automake wakeup
Sometimes, when the max thread count is exceeded, it is possible for
the scheduler to lock up. This is because the scheduler is only run when
a new job is put into the queue, and so it is possible for a job to be
blocked by thread limitations and be left sitting in the queue. If the
main program is waiting on this job to finish before scheduling a new one,
then the system deadlocks.

I added a second background thread to the scheduler that wakes the
scheduler up every 20us to resolve this and prevent these deadlocks.